fix browserify
