+++ src/test_async_backends.ts
@@ -75,5 +75,6 @@ setTestEnvs([{
 
 const runner = new jasmine();
 // exclude worker test
-runner.loadConfig({spec_files: ['dist/**/**_test.js'], random: false});
+runner.loadConfig({spec_files: ['dist/**/**_test.js',
+    'dist/worker_test_node.js'], random: false});
 runner.execute();
